It works!!!! Sorta

+++ BlockByBlock/net/minecraft/src/NBTTagEnd.cs
@@ -19,7 +19,7 @@
 		{
 			get
 			{
-				return (sbyte)0;
+				return 0;
 			}
 		}
